cns/middlewares/k8sSwiftV2_windows.go
package middlewares

import (
	"github.com/Azure/azure-container-networking/cns"
	"github.com/Azure/azure-container-networking/cns/logger"
)

// for AKS L1VH, do not set default route on infraNIC to avoid customer pod reaching all infra vnet services
// default route is set for secondary interface NIC(i.e,delegatedNIC)
func (k *K8sSWIFTv2Middleware) setRoutes(podIPInfo *cns.PodIpInfo) error {
	if podIPInfo.NICType == cns.InfraNIC {
		logger.Printf("[SWIFTv2Middleware] skip setting default route on InfraNIC interface")
		podIPInfo.SkipDefaultRoutes = true
	}
	return nil
}
